LPD-83226 Skip items without delete permission in bulk delete action

+++ modules/apps/site/site-cms-site-initializer/src/main/java/com/liferay/site/cms/site/initializer/internal/bulk/selection/DeleteObjectBulkSelectionAction.java
@@ -13,8 +13,12 @@ import com.liferay.object.rest.manager.v1_0.DefaultObjectEntryManager;
 import com.liferay.object.rest.manager.v1_0.DefaultObjectEntryManagerProvider;
 import com.liferay.object.rest.manager.v1_0.ObjectEntryManagerRegistry;
 import com.liferay.object.service.ObjectEntryFolderService;
+import com.liferay.object.service.ObjectEntryService;
 import com.liferay.portal.kernel.feature.flag.FeatureFlagManagerUtil;
 import com.liferay.portal.kernel.model.User;
+import com.liferay.portal.kernel.security.permission.ActionKeys;
+import com.liferay.portal.kernel.security.permission.PermissionThreadLocal;
+import com.liferay.portal.kernel.security.permission.resource.ModelResourcePermission;
 import com.liferay.portal.kernel.service.ServiceContext;
 import com.liferay.site.cms.site.initializer.bulk.selection.BaseObjectBulkSelectionAction;
 import com.liferay.trash.TrashHelper;
@@ -48,19 +52,34 @@ public class DeleteObjectBulkSelectionAction
 				objectDefinitionLocalService.getObjectDefinition(
 					objectObjectEntry.getObjectDefinitionId());
 
-			DefaultObjectEntryManager defaultObjectEntryManager =
-				DefaultObjectEntryManagerProvider.provide(
-					_objectEntryManagerRegistry.getObjectEntryManager(
-						objectDefinition.getCompanyId(),
-						objectDefinition.getStorageType()));
+			ModelResourcePermission<ObjectEntry> modelResourcePermission =
+				_objectEntryService.getModelResourcePermission(
+					objectDefinition.getObjectDefinitionId());
 
-			defaultObjectEntryManager.deleteObjectEntry(
-				objectDefinition, objectObjectEntry.getObjectEntryId());
+			if (modelResourcePermission.contains(
+					PermissionThreadLocal.getPermissionChecker(),
+					objectObjectEntry, ActionKeys.DELETE)) {
+
+				DefaultObjectEntryManager defaultObjectEntryManager =
+					DefaultObjectEntryManagerProvider.provide(
+						_objectEntryManagerRegistry.getObjectEntryManager(
+							objectDefinition.getCompanyId(),
+							objectDefinition.getStorageType()));
+
+				defaultObjectEntryManager.deleteObjectEntry(
+					objectDefinition, objectObjectEntry.getObjectEntryId());
+			}
 		}
 		else {
 			ObjectEntryFolder objectEntryFolder = (ObjectEntryFolder)object;
 
-			_deleteObjectEntryFolder(objectEntryFolder);
+			if (_objectEntryFolderModelResourcePermission.contains(
+					PermissionThreadLocal.getPermissionChecker(),
+					objectEntryFolder.getObjectEntryFolderId(),
+					ActionKeys.DELETE)) {
+
+				_deleteObjectEntryFolder(objectEntryFolder);
+			}
 		}
 	}
 
@@ -80,11 +99,20 @@ public class DeleteObjectBulkSelectionAction
 		}
 	}
 
+	@Reference(
+		target = "(model.class.name=com.liferay.object.model.ObjectEntryFolder)"
+	)
+	private ModelResourcePermission<ObjectEntryFolder>
+		_objectEntryFolderModelResourcePermission;
+
 	@Reference
 	private ObjectEntryFolderService _objectEntryFolderService;
 
 	@Reference
 	private ObjectEntryManagerRegistry _objectEntryManagerRegistry;
+
+	@Reference
+	private ObjectEntryService _objectEntryService;
 
 	@Reference
 	private TrashHelper _trashHelper;
